@@ -1,10 +1,11 @@
 {
     "name": "formr/formr",
-    "description": "Formr is a PHP micro-framework which helps you build and validate forms quickly, painlessly, and without all the messy overhead.",
+    "description": "Formr is a PHP library which helps you build and validate forms quickly, painlessly, and without all the messy overhead.",
     "type": "library",
     "license": "GPL-2.0-only",
     "require": {
-        "php": ">=5.6"
+        "php": ">=8.1",
+        "ext-mbstring": "*"
     },
     "authors": [
         {
