@@ -4,8 +4,8 @@ trait Bulma
 {
     # Wrapper for the Bulma framework
     # https://bulma.io
-    
-    public static function bulma_css($key = '')
+
+    public static function bulma_css($key = ''): array|string
     {
         $array = [
             'alert-e' => 'is-danger',
@@ -16,197 +16,102 @@ trait Bulma
             'button-danger' => 'button is-danger',
             'button-primary' => 'button is-primary',
             'button-secondary' => 'button is-link',
-            'checkbox' => 'checkbox',
-            'div' => 'div',
+            'checkbox' => '',
+            'div' => 'my-3',
             'file' => 'file-input',
             'help' => 'help',
             'is-invalid' => 'is-danger',
             'is-valid' => 'is-success',
             'input' => 'input',
             'label' => 'label',
-            'radio' => 'radio',
+            'radio' => '',
             'textarea' => 'textarea',
         ];
-        
+
         if ($key) {
-            return $array[$key];
-        } else {
-            return $array;
+            return $array[$key] ?? '';
         }
+
+        return $array;
     }
 
-    public function bulma($element = '', $data = '')
+    public function bulma($element = '', $data = ''): string
     {
         if (empty($data)) {
-            return false;
+            return '';
         }
-        
-        # define our $return variable with a new line
-        $return = $this->nl;
-        
+
+        # define our $return variable
+        $return = null;
+
         # add a comment if $form->comments is enabled
-        $return .= $this->formr->_print_field_comment($data);
-        
-        # build a checkbox
-        if ($data['type'] == 'checkbox' || $data['type'] == 'radio')
-        {
-            # open the wrapper
-            if ($data['type'] == 'checkbox' && $this->formr->use_element_wrapper_div) {
-                $return .= '<div class="field">' . $this->nl;
-            }
-            
-            $return .= '<label class="' . $data['type'] . '">' . $this->nl;
-            $return .=     $element . $this->nl;
-            $return .=     $data['label'] . $this->nl;
-            $return .= '</label>' . $this->nl;
-            
-            # show error message
-            if ($this->formr->submitted() && $this->formr->in_errors($data['name']) && $this->formr->inline_errors) {
-                $return .= '<p class="help is-danger">' . $this->formr->errors[$data['name']] . '</p>';
-            }
-            
-            # close the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= '</div>' . $this->nl;
-            }
-            
-            return $return;
-            
+        $return .= $this->_print_field_comment($data);
+
+        # open the wrapper
+        if ($this->formr->use_element_wrapper_div) {
+            $return .= "<div class=\"field\">".PHP_EOL;
         }
-        elseif ($data['type'] == 'file') {
-            
-            # file element
-            
-            # open the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= '<div class="field">' . $this->nl;
-            }
-            
-            if ($this->formr->is_not_empty($data['label'])) {
-                $return .= '<label class="label" for="' . $this->formr->make_id($data) . '">' . $this->nl;
-                $return .= "\t" . $data['label'];
-                $return .= $this->formr->insert_required_indicator($data) . $this->nl;
-                $return .= '</label>' . $this->nl;
-            }
-            
-            $return .= '  <div class="file">' . $this->nl;
-            $return .= '    <label class="file-label">' . $this->nl;
-            $return .=       $element . $this->nl;
-            $return .= '      <span class="file-cta">' . $this->nl;
-            $return .= '        <span class="file-icon">' . $this->nl;
-            $return .= '          <i class="fas fa-upload"></i>' . $this->nl;
-            $return .= '        </span>' . $this->nl;
-            $return .= '        <span class="file-label">' . $this->nl;
-            $return .= '          Choose file(s)...' . $this->nl;
-            $return .= '        </span>' . $this->nl;
-            $return .= '      </span>' . $this->nl;
-            $return .= '    </label>' . $this->nl;
-            $return .= '  </div>' . $this->nl;
-            
-            # show error message
-            if ($this->formr->submitted() && $this->formr->in_errors($data['name']) && $this->formr->inline_errors) {
-                $return .= '<p class="help is-danger">' . $this->formr->errors[$data['name']] . '</p>';
-            }
-            
-            # close the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= '</div>' . $this->nl;
-            }
-            
-            return $return;
-            
-        }
-        elseif ($data['type'] == 'select') {
-            
-            #select menu
-            
-            # open the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= '<div class="field">' . $this->nl;
-            }
-            
-            if ($this->formr->is_not_empty($data['label'])) {
-                $return .= '<label class="label" for="' . $this->formr->make_id($data) . '">' . $this->nl;
-                $return .= "\t" . $data['label'];
-                $return .= $this->formr->insert_required_indicator($data) . $this->nl;
-                $return .= '</label>' . $this->nl;
-            }
-            
-            $return .= '  <div class="control">' . $this->nl;
-            $return .= '    <div class="select">' . $this->nl;
-            $return .=       $element . $this->nl;
-            $return .= '    </div>' . $this->nl;
-            $return .= '  </div>' . $this->nl;
-            
-            # show error message
-            if ($this->formr->submitted() && $this->formr->in_errors($data['name']) && $this->formr->inline_errors) {
-                $return .= '<p class="help is-danger">' . $this->formr->errors[$data['name']] . '</p>';
-            }
-            
-            # close the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= '</div>' . $this->nl;
-            }
-            
-            return $return;
-            
+
+        if ($this->type_is_checkbox($data)) {
+            # checkbox and radio
+            $return .= "<label class=\"{$data['type']}\">".PHP_EOL;
+            $return .= $element.PHP_EOL;
+            $return .= $data['label'].PHP_EOL;
+            $return .= "</label>".PHP_EOL;
+            $return .= $this->formr->printWrapperMessages($data);
+        } elseif ($data['type'] == 'file') {
+            # files
+            $return .= $this->formr->printWrapperLabel($data).PHP_EOL;
+            $return .= "<div class=\"file\">".PHP_EOL;
+            $return .= "    <label class=\"file-label\">".PHP_EOL;
+            $return .=          $element.PHP_EOL;
+            $return .= "        <span class=\"file-cta\">".PHP_EOL;
+            $return .= "            <span class=\"file-icon\">".PHP_EOL;
+            $return .= "                <i class=\"fas fa-upload\"></i>".PHP_EOL;
+            $return .= "            </span>".PHP_EOL;
+            $return .= "            <span class=\"file-label\">".PHP_EOL;
+            $return .= "                Choose file(s)...".PHP_EOL;
+            $return .= "            </span>".PHP_EOL;
+            $return .= "        </span>".PHP_EOL;
+            $return .= "    </label>".PHP_EOL;
+            $return .= "</div>".PHP_EOL;
+        } elseif ($data['type'] == 'select') {
+            # select menus
+            $return .= $this->formr->printWrapperLabel($data).PHP_EOL;
+            $return .= "<div class=\"control\">".PHP_EOL;
+            $return .= "    <div class=\"select\">".PHP_EOL;
+            $return .=          $element.PHP_EOL;
+            $return .= "    </div>".PHP_EOL;
+            $return .=      $this->formr->printWrapperMessages($data);
+            $return .= "</div>".PHP_EOL;
         } else {
-        
             # everything else
-            
-            # open the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= $this->nl . '<div id="_' . $this->formr->make_id($data) . '" class="field">' . $this->nl;
-            }
-            
-            if ($this->formr->is_not_empty($data['label'])) {
-                $return .= '<label class="label" for="' . $this->formr->make_id($data) . '">' . $this->nl;
-                $return .= "\t" . $data['label'];
-                $return .= $this->formr->insert_required_indicator($data) . $this->nl;
-                $return .= '</label>' . $this->nl;
-            }
-            
-            $return .= '<div class="control has-icons-right">' . $this->nl;
-            $return .= "\t" . $element . $this->nl;
-            
-            # show fontawesome icons and highlight fields if error or success
+            $return .= $this->formr->printWrapperLabel($data).PHP_EOL;
+            $return .= "<div class=\"control has-icons-left has-icons-right\">".PHP_EOL;
+            $return .= $element.PHP_EOL;
             if ($this->formr->submitted()) {
+                # show fontawesome icons and highlight fields if error or success
                 if ($this->formr->in_errors($data['name'])) {
-                    $return .= '<span class="icon is-small is-right">' . $this->nl;
-                    $return .= '    <i class="fas fa-exclamation-triangle"></i>' . $this->nl;
-                    $return .= '</span>' . $this->nl;
+                    $return .= '<span class="icon is-small is-right">'.PHP_EOL;
+                    $return .= '    <i class="fas fa-exclamation-triangle"></i>'.PHP_EOL;
+                    $return .= '</span>'.PHP_EOL;
                 } else {
-                    if ($this->formr->show_valid && !in_array($data['type'], $this->formr->excluded_types)) {
-                        $return .= '<span class="icon is-small is-right">' . $this->nl;
-                        $return .= '    <i class="fas fa-check"></i>' . $this->nl;
-                        $return .= '</span>' . $this->nl;
+                    if ($this->formr->show_valid && ! in_array($data['type'], $this->formr->excluded_types)) {
+                        $return .= '<span class="icon is-small is-right">'.PHP_EOL;
+                        $return .= '    <i class="fas fa-check"></i>'.PHP_EOL;
+                        $return .= '</span>'.PHP_EOL;
                     }
                 }
             }
-            
-            $return .= '</div>' . $this->nl;
-            
-            # show inline help or error message
-            if (!empty($data['inline'])) {
-                if ($this->formr->is_in_brackets($data['inline'])) {
-                    if ($this->formr->in_errors($data['name'])) {
-                        $return .= '<p class="help is-danger">' . trim($data['inline'], '[]') . '</p>' . $this->nl;
-                    }
-                } else {
-                    $return .= '<p class="help">' . $data['inline'] . '</p>' . $this->nl;
-                }
-            } else {
-                if ($this->formr->submitted() && $this->formr->in_errors($data['name']) && $this->formr->inline_errors) {
-                    $return .= '<p class="help is-danger">' . $this->formr->errors[$data['name']] . '</p>';
-                }
-            }
-            
-            # close the wrapper
-            if ($this->formr->use_element_wrapper_div) {
-                $return .= '</div>' . $this->nl;
-            }
-            
-            return $return . $this->nl;
+            $return .= $this->formr->printWrapperMessages($data);
+            $return .= "</div>".PHP_EOL;
         }
+
+        # close the wrapper
+        if ($this->formr->use_element_wrapper_div) {
+            $return .= '</div>'.PHP_EOL;
+        }
+
+        return $return.PHP_EOL;
     }
 }
